+++ src/shared/components/palette/case-flag/domain/case-flag.model.ts
@@ -52,5 +52,7 @@ export interface FlagsWithFormGroupPath {
 export interface FlagDetailDisplayWithFormGroupPath {
   flagDetailDisplay: FlagDetailDisplay;
   pathToFlagsFormGroup: string;
-  caseField: CaseField
+  caseField: CaseField;
+  roleOnCase?: string;
+  label?: string;
 }